+++ src/main/kotlin/dev/aaronhowser/mods/excessive_utilities/block_entity/transfer_node/ItemTransferNodeBlockEntity.kt
@@ -21,7 +21,6 @@ import net.minecraft.world.entity.item.ItemEntity
 import net.minecraft.world.entity.player.Inventory
 import net.minecraft.world.entity.player.Player
 import net.minecraft.world.inventory.AbstractContainerMenu
-import net.minecraft.world.inventory.ContainerData
 import net.minecraft.world.item.ItemStack
 import net.minecraft.world.level.block.state.BlockState
 import net.minecraft.world.phys.AABB
@@ -57,58 +56,9 @@ class ItemTransferNodeBlockEntity(
 		return level.getCapability(Capabilities.ItemHandler.BLOCK, placedOnPos, placedOnDirection.opposite)
 	}
 
-	// Pull from distant inventories into the buffer,
-	// then push from the buffer into the parent inventory.
-	// if there's already stuff in the buffer, there's no reason to try to continue filling it,
-	// so just reset the ping and don't continue searching
-	override fun pullerTick(level: ServerLevel) {
-		pushIntoParent(level)
+	override fun getBufferAmount(): Int = bufferContainer.getItem(0).count
 
-		if (!bufferContainer.isEmpty) {
-			if (!hasPseudoRoundRobinUpgrade()) {
-				ping.reset()
-			}
-			return
-		}
-
-		// Try to pull from inventories at the ping position
-		// If nothing was pulled, move the ping forward and try again next tick
-		// If something was pulled, keep the ping where it is so it can continue pulling from it next tick
-
-		pullFromPingPos(level)
-
-		if (bufferContainer.isEmpty || hasPseudoRoundRobinUpgrade()) {
-			ping.march(level)
-		}
-	}
-
-	// Pull from the parent inventory into the buffer,
-	// then search for somewhere to push the items in the buffer to.
-	// If there's nothing in the buffer, don't bother searching for somewhere to put it
-	override fun pusherTick(level: ServerLevel) {
-		pullFromParent(level)
-
-		if (bufferContainer.isEmpty) {
-			if (!hasPseudoRoundRobinUpgrade()) {
-				ping.reset()
-			}
-			return
-		}
-
-		// Try to push into inventories at the ping position
-		// If nothing was pushed, move the ping forward and try again next tick
-		// If something was pushed, keep the ping where it is so it can continue pushing into it next tick
-
-		val amountBefore = bufferContainer.getItem(0).count
-		pushIntoPingPos(level)
-		val amountAfter = bufferContainer.getItem(0).count
-
-		if (amountBefore == amountAfter || hasPseudoRoundRobinUpgrade()) {
-			ping.march(level)
-		}
-	}
-
-	private fun pushIntoPingPos(level: ServerLevel) {
+	override fun pushIntoPingPos(level: ServerLevel) {
 		val stackInBuffer = bufferContainer.getItem(0)
 		if (stackInBuffer.isEmpty) return
 
@@ -132,7 +82,7 @@ class ItemTransferNodeBlockEntity(
 		}
 	}
 
-	private fun pullFromPingPos(level: ServerLevel) {
+	override fun pullFromPingPos(level: ServerLevel) {
 		val itemHandlers = getItemHandlersAroundPing(level)
 		if (itemHandlers.isEmpty()) return
 
@@ -152,20 +102,12 @@ class ItemTransferNodeBlockEntity(
 	}
 
 	private fun getItemHandlersAroundPing(level: ServerLevel): List<IItemHandler> {
-		val possibleDirections = ping.getNextDirections(level)
-		val pingPos = ping.currentPingPos
-
-		val handlers = mutableListOf<IItemHandler>()
-		for (dir in possibleDirections) {
-			val neighborPos = pingPos.relative(dir)
-			val handler = level.getCapability(Capabilities.ItemHandler.BLOCK, neighborPos, dir.opposite) ?: continue
-			handlers.add(handler)
+		return getCapabilitiesAroundPing(level) { neighborPos, side ->
+			level.getCapability(Capabilities.ItemHandler.BLOCK, neighborPos, side)
 		}
-
-		return handlers
 	}
 
-	private fun pushIntoParent(level: ServerLevel) {
+	override fun pushIntoParent(level: ServerLevel) {
 		val parentHandler = getParentItemHandler(level) ?: return
 
 		val stackInBuffer = bufferContainer.getItem(0)
@@ -287,7 +229,7 @@ class ItemTransferNodeBlockEntity(
 		return true
 	}
 
-	private fun pullFromParent(level: ServerLevel) {
+	override fun pullFromParent(level: ServerLevel) {
 		if (worldInteraction(level)) return
 
 		val parentHandler = getParentItemHandler(level) ?: return
@@ -325,24 +267,6 @@ class ItemTransferNodeBlockEntity(
 			break
 		}
 	}
-
-	private val containerData =
-		object : ContainerData {
-			override fun getCount(): Int = CONTAINER_DATA_SIZE
-
-			override fun get(index: Int): Int {
-				return when (index) {
-					X_DATA_INDEX -> ping.currentPingPos.x
-					Y_DATA_INDEX -> ping.currentPingPos.y
-					Z_DATA_INDEX -> ping.currentPingPos.z
-					else -> 0
-				}
-			}
-
-			override fun set(index: Int, value: Int) {
-				// Noop
-			}
-		}
 
 	override fun createMenu(containerId: Int, playerInventory: Inventory, player: Player): AbstractContainerMenu {
 		return ItemTransferNodeMenu(
@@ -383,11 +307,6 @@ class ItemTransferNodeBlockEntity(
 
 		const val BUFFER_CONTAINER_SIZE = 1
 		const val FILTER_CONTAINER_SIZE = 1
-
-		const val CONTAINER_DATA_SIZE = 3
-		const val X_DATA_INDEX = 0
-		const val Y_DATA_INDEX = 1
-		const val Z_DATA_INDEX = 2
 
 	}